// angular-sample/angular-webapp/client/app/root-module/models/company-site-view.ts
export class CompanySiteView {
    Id: string;
    Name: string;
    ParentId: string;
    ParentCompanyId: string;
    RootId: string;
    IsSite: boolean;
    CreatedOn: Date;
    ModifiedOn: Date;
    CreatedBy: string;
    ModifiedBy: string;
    IsDeleted: boolean;
    LCid: number;
    Version: string;
}